* 'dev-deploy' of https://github.com/WeBankFinTech/WeBASE-Node-Manager: (131 commits)
  v1.3.1
  update fastjson
  fix index out range
  v1.3.1
  revert group async rename
  rename map service
  fix all node invalid
  stop group manually set map as invalid
  add schedule threads to 50
  add log of update map
  adjust async task
  add update front internal
  fix catch 500 and update front
  catch access exception of front
  add status front
  add front status
  adjust error code
  add batch operate status list
  fix null valid
  fix front_group_map by local groupList
  ...

# Conflicts:
#	script/webase-ddl.sql
#	src/main/java/com/webank/webase/node/mgr/front/FrontMapper.java
#	src/main/java/com/webank/webase/node/mgr/front/FrontService.java
#	src/main/java/com/webank/webase/node/mgr/front/entity/TbFront.java
#	src/main/java/com/webank/webase/node/mgr/frontgroupmap/FrontGroupMapService.java
#	src/main/java/com/webank/webase/node/mgr/frontgroupmap/entity/TbFrontGroupMap.java
#	src/main/java/com/webank/webase/node/mgr/group/GroupService.java
#	src/main/java/com/webank/webase/node/mgr/group/entity/TbGroup.java
#	src/main/java/com/webank/webase/node/mgr/node/NodeService.java
#	src/main/resources/application.yml
#	src/main/resources/mapper/FrontMapper.xml
#	src/main/resources/mapper/GroupMapper.xml

+++ src/main/java/com/webank/webase/node/mgr/base/enums/FrontStatusEnum.java
@@ -23,12 +23,12 @@ import lombok.ToString;
 @ToString
 @AllArgsConstructor
 public enum FrontStatusEnum {
-    INITIALIZED((byte) 0, "initialized"),
-    RUNNING((byte) 1, "running"),
-    STOPPED((byte) 2, "stopped"),
+    INITIALIZED( 0, "initialized"),
+    RUNNING( 1, "running"),
+    STOPPED(2, "stopped"),
     ;
 
-    private byte id;
+    private int id;
     private String description;
 
 
@@ -37,7 +37,7 @@ public enum FrontStatusEnum {
      * @param id
      * @return
      */
-    public static FrontStatusEnum getById(byte id) {
+    public static FrontStatusEnum getById(int id) {
         for (FrontStatusEnum value : FrontStatusEnum.values()) {
             if (value.id == id) {
                 return value;
